refactor: hooks 폴더 분리
[#360]
