+++ frontend/src/hooks/google-maps/useCalculatedMapDelta.ts
