+++ frontend/src/hooks/google-maps/useCurrentPosition.ts
